requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
Pillow>=8.2.0
nltk>=3.6.2
beautifulsoup4>=4.9.3
Flask>=2.0.1
dvc>=2.5.4
mlflow>=1.18.0
python-dotenv>=0.19.0
torchmetrics>=0.5.0
tqdm>=4.61.1
psutil>=5.8.0
pyyaml>=5.4.1
requests>=2.25.1
click>=8.0.1
watchdog>=2.1.3
joblib>=1.0.1
setuptools>=57.0.0
wheel>=0.36.2
certifi>=2021.5.30
rich==13.5.2